Move SignedDuration to mijia crate.
